requirements.txt
dj-database-url==0.4.1
Django==1.10.3
django-el-pagination==3.0.1
gunicorn==19.6.0
httplib2==0.9.2
psycopg2==2.6.2
python-dateutil==2.6.0
pytz==2016.7
requests==2.12.1
simplejson==3.10.0
six==1.10.0
whitenoise==3.2.2
